+++ frontend/src/mocks/handlers/studyInfoHandlers.ts
@@ -1,7 +1,7 @@
 import { rest } from 'msw';
 
-export const checkParticipantCodeHandlers = [
-  rest.get('api/studies', async (req, res, ctx) => {
+export const studyInfoHandlers = [
+  rest.get('/api/studies', async (req, res, ctx) => {
     const accessToken =
       'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxNjkxNTY4NDI4LCJleHAiOjE2OTE1NzIwMjh9.BfGH7jBxO_iixmlpzxHKV7d9ekJPegLxrpY9ME066ro';
     const requestAuthToken = req.headers.get('Authorization')?.split(' ')[1];
@@ -88,6 +88,54 @@ export const checkParticipantCodeHandlers = [
       //   message: '에러',
       //   code: 1234,
       // }),
+      ctx.delay(1000),
+    );
+  }),
+
+  rest.post('/api/studies', (req, res, ctx) => {
+    const accessToken =
+      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxNjkxNTY4NDI4LCJleHAiOjE2OTE1NzIwMjh9.BfGH7jBxO_iixmlpzxHKV7d9ekJPegLxrpY9ME066ro';
+
+    const requestAuthToken = req.headers.get('Authorization')?.split(' ')[1];
+
+    const newAccessToken =
+      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxMjM0NTY3fQ.NUiutjXo0mcIBU5fWxfjpBEvPxakFiBaUCg4THKAYpQ';
+
+    if (requestAuthToken === newAccessToken)
+      return res(
+        // 성공
+        ctx.status(201),
+        ctx.set({ 'Content-Type': 'application/json', Location: '/api/studies/1' }),
+
+        // 에러
+        // ctx.status(404),
+        // ctx.json({
+        //   code: 1201,
+        //   message: '에러',
+        // }),
+
+        ctx.delay(1000),
+      );
+
+    if (accessToken !== requestAuthToken)
+      return res(
+        ctx.status(401),
+        ctx.json({ message: '유효하지 않은 엑세스 토큰입니다.', code: 1403 }),
+        ctx.delay(1000),
+      );
+
+    return res(
+      // 성공
+      ctx.status(201),
+      ctx.set({ 'Content-Type': 'application/json', Location: '/api/studies/1' }),
+
+      // 에러
+      // ctx.status(404),
+      // ctx.json({
+      //   code: 1201,
+      //   message: '에러',
+      // }),
+
       ctx.delay(1000),
     );
   }),
